+++ Data/BuildVersionData/Shared.lua
@@ -1,0 +1,4 @@
+local _, addon = ...;
+addon.Data.BuildVersionData = {};
+addon.Data.BuildVersionData.Shared = {};
+local shared = addon.Data.BuildVersionData.Shared;