+++ Data/EventData/00_Mainline/08_BattleForAzeroth.lua
@@ -1,0 +1,138 @@
+local _, addon = ...;
+local shared = addon.Data.EventData.Shared;
+local type = addon.Objects.EventType;
+
+KrowiAF.EventData.BattleForAzeroth = {
+    { -- Faction Assaults: Tiragarde Sound
+        5896, type.World,
+        935,
+        2065630,
+		{addon.L["Faction Assaults"], addon.L["Tiragarde Sound"]},
+        addon.L["Faction Assaults"],
+        1,
+        7,
+        false,
+        876
+    },
+    { -- Faction Assaults: Drustvar
+        5964, type.World,
+        936,
+        2065567,
+		{addon.L["Faction Assaults"], addon.L["Drustvar"]},
+        addon.L["Faction Assaults"],
+        1,
+        7,
+        false,
+        876
+    },
+    { -- Faction Assaults: Stormsong Valley
+        5966, type.World,
+        937,
+        2065627,
+		{addon.L["Faction Assaults"], addon.L["Stormsong Valley"]},
+        addon.L["Faction Assaults"],
+        1,
+        7,
+        false,
+        876
+    },
+    { -- Faction Assaults: Nazmir
+        5969, type.World,
+        938,
+        2032229,
+		{addon.L["Faction Assaults"], addon.L["Nazmir"]},
+        addon.L["Faction Assaults"],
+        1,
+        7,
+        false,
+        875
+    },
+    { -- Faction Assaults: Vol'dun
+        5970, type.World,
+        939,
+        2065632,
+		{addon.L["Faction Assaults"], addon.L["Vol'dun"]},
+        addon.L["Faction Assaults"],
+        1,
+        7,
+        false,
+        875
+    },
+    { -- Faction Assaults: Zuldazar
+        5973, type.World,
+        934,
+        2065640,
+		{addon.L["Faction Assaults"], addon.L["Zuldazar"]},
+        addon.L["Faction Assaults"],
+        1,
+        7,
+        false,
+        875
+    },
+    { -- Defending Azeroth Assaults: N'Zoth (Uldum)
+        6486, type.World,
+        942,
+        3196264,
+		{addon.L["Defending Azeroth Assaults"], addon.L["N'Zoth (Uldum)"]},
+        addon.L["Defending Azeroth Assaults"],
+        2,
+        7,
+        false,
+        12
+    },
+    { -- Defending Azeroth Assaults: Amathet
+        6487, type.World,
+        942,
+        415054,
+		{addon.L["Defending Azeroth Assaults"], addon.L["Amathet"]},
+        addon.L["Defending Azeroth Assaults"],
+        2,
+        7,
+        false,
+        12
+    },
+    { -- Defending Azeroth Assaults: Aqir
+        6488, type.World,
+        942,
+        458247,
+		{addon.L["Defending Azeroth Assaults"], addon.L["Aqir"]},
+        addon.L["Defending Azeroth Assaults"],
+        2,
+        7,
+        false,
+        12
+    },
+    { -- Defending Azeroth Assaults: N'Zoth (Vale of Eternal Blossoms)
+        6489, type.World,
+        941,
+        3196265,
+		{addon.L["Defending Azeroth Assaults"], addon.L["N'Zoth (Vale of Eternal Blossoms)"]},
+        addon.L["Defending Azeroth Assaults"],
+        2,
+        7,
+        false,
+        424
+    },
+    { -- Defending Azeroth Assaults: Mogu
+        6490, type.World,
+        941,
+        801016,
+		{addon.L["Defending Azeroth Assaults"], addon.L["Mogu"]},
+        addon.L["Defending Azeroth Assaults"],
+        2,
+        7,
+        false,
+        424
+    },
+    { -- Defending Azeroth Assaults: Mantid
+        6491, type.World,
+        941,
+        615301,
+		{addon.L["Defending Azeroth Assaults"], addon.L["Mantid"]},
+        addon.L["Defending Azeroth Assaults"],
+        2,
+        7,
+        false,
+        424
+    },
+};