+++ Data/EventData/Shared.lua
@@ -1,0 +1,4 @@
+local _, addon = ...;
+addon.Data.EventData = {};
+addon.Data.EventData.Shared = {};
+local shared = addon.Data.EventData.Shared;