+++ Localization/enUS.lua
@@ -7,7 +7,7 @@ addon.Plugins:LoadLocalization(L);
 -- [[ https://legacy.curseforge.com/wow/addons/krowi-achievement-filter/localization ]] --
 -- [[ Everything after this line is automatically generated from CurseForge and is not meant for manual edit - SOURCETOKEN - AUTOGENTOKEN ]] --
 
--- [[ Exported at 2024-07-29 22-11-27 ]] --
+-- [[ Exported at 2024-10-24 17-34-18 ]] --
 L["%c"] = true
 L["%d/%m/%Y %I:%M %p"] = true
 L["%m/%d/%Y %I:%M %p"] = true
@@ -256,6 +256,7 @@ L["Down"] = true
 L["Dragon Glyphs"] = true
 L["Dragon Racing"] = true
 L["Dragonbane Keep"] = true
+L["Dragonblight"] = true
 L["Dragonflight (pre-patch)"] = true
 L["Dragonriding"] = true
 L["Dragonriding Races"] = true
@@ -265,6 +266,7 @@ L["Drops"] = true
 L["Drustvar"] = true
 L["Dueler's Guild"] = true
 L["during"] = true
+L["Dustwallow Marsh"] = true
 L["Earned By"] = true
 L["Earned By Desc"] = [=[Header
 |T:1:8|t- {achievementPointsEarnedBy}: Display {achievementPointsEarnedBy_wc} your current character and/or {achievementPointsEarnedBy_wc} your account ({achievementPointsEarnedByFormat}) (1).
@@ -466,6 +468,7 @@ L["Force two columns"] = true
 L["Force two columns Desc"] = "Enable / disable the algorithm to force two achievement objectives columns. When enabled, if only one column of objectives would be shown due to the text of one or more objectives being too long, a second column will be forced based on the other settings."
 L["Force two columns threshold"] = true
 L["Force two columns threshold Desc"] = "The number of criteria the achievement must have before we force a second column if normally only one would be shown."
+L["Forget Me Not: Next"] = true
 L["Format"] = true
 L["Fracture Chambers"] = true
 L["Fractures in Time"] = true
@@ -705,6 +708,7 @@ L["Plugins"] = true
 L["Plunderstorm"] = true
 L["Points"] = true
 L["Pop ups"] = true
+L["Pre-Patch Event"] = true
 L["Presets"] = true
 L["Primal Storms"] = true
 L["Primalist Tomorrow"] = true
@@ -729,6 +733,7 @@ Options (1):
 |T:1:8|t- {numSearchPreviews} ({numSearchPreviewsNumber})
 
 (1) Options can be found under {gameMenu} {arrow} {interface} {arrow} {addOns} {arrow} {addonName}]=]
+L["Radiant Echoes"] = true
 L["Rage of the Firelands"] = true
 L["Rated"] = true
 L["Ravine River Run"] = true
@@ -809,6 +814,7 @@ L["Search options"] = true
 L["Search preview"] = true
 L["Search Results"] = true
 L["Search Rewards"] = true
+L["Searing Gorge"] = true
 L["Season"] = true
 L["Second column threshold"] = true
 L["Second column threshold Desc"] = "Achievements with more objectives than this number will be visible in two columns instead of one."
@@ -943,6 +949,7 @@ L["Spreading the Light"] = true
 L["Stables"] = true
 L["StarCraft II"] = true
 L["Start Time"] = true
+L["Stop tracking"] = true
 L["Stormheim"] = true
 L["Stormsong Valley"] = true
 L["Strand of the Ancients"] = true
@@ -975,15 +982,19 @@ L["The Azure Span Sprint"] = true
 L["The Big Dig: Active"] = true
 L["The Big Dig: Next"] = true
 L["The Big Dig: Soon"] = true
+L["The Cruelty of Dornic: Next"] = true
 L["The Entitled"] = true
 L["the event name with run time"] = "the {eventName} from {startDate} until {endDate}"
 L["The Flowing Forest Flight"] = true
 L["The Iron Tide"] = true
+L["The Rivals: Next"] = true
 L["The Shattering"] = true
+L["The Thraegar's Descent: Next"] = true
 L["The Thunder King"] = true
 L["The Tomb of Sargeras"] = true
 L["The Vakthros Ascent"] = true
 L["The Waking Shore"] = true
+L["The Wanderer: Next"] = true
 L["The War Within (pre-patch)"] = true
 L["Theater Troupe"] = true
 L["This achievement"] = true