+++ Objects/BuildVersion.lua
@@ -5,11 +5,17 @@ objects.BuildVersion = {};
 local buildVersion = objects.BuildVersion;
 
 buildVersion.__index = buildVersion;
-function buildVersion:New(id, name, value, description)
+function buildVersion:New(id, major, minor, patch, name, description)
     local instance = setmetatable({}, buildVersion);
     instance.Id = id or 0;
+    instance.Major = major or 0;
+    instance.Minor = minor or 0;
+    instance.Patch = patch or 0;
     instance.Name = name;
-    instance.Value = value;
     instance.Description = description;
     return instance;
+end
+
+function buildVersion:SetInUse()
+    self.InUse = true;
 end